core: stdcm: regroup all simulation-related methods into a new class
