+++ core/src/main/java/fr/sncf/osrd/api/stdcm/graph/STDCMEdgeBuilder.java
@@ -120,7 +120,7 @@ public class STDCMEdgeBuilder {
     /** Creates all edges that can be accessed on the given route, using all the parameters specified. */
     public Collection<STDCMEdge> makeAllEdges() {
         if (envelope == null)
-            envelope = STDCMUtils.simulateRoute(
+            envelope = STDCMSimulations.simulateRoute(
                     route,
                     startSpeed,
                     startOffset,
